add bin packing solution

+++ models/cli.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel, Field
 class Cli(BaseModel):
     input: str = Field(default="data/input.csv")
     output: str = Field(default="data/output.csv")
-    sum: int
+    sum: float
